# minor performance improvements to url_parse (#429)