--- a/R/parse.R
+++ b/R/parse.R
@@ -1,7 +1,7 @@
 parse_media <- function(x) {
   # https://datatracker.ietf.org/doc/html/rfc2616#section-3.7
   pieces <- parse_delim(x, ";")
-  params <- parse_name_equals_value(pieces[-1])
+  params <- parse_name_equals_value(pieces[-1], ";")
 
   if (is_empty(pieces)) {
     list(type = NA_character_)
@@ -15,7 +15,7 @@ parse_www_authenticate <- function(x) {
   # https://stackoverflow.com/questions/10239970
 
   pieces <- parse_in_half(x, " ")
-  params <- parse_name_equals_value(parse_delim(pieces[[2]], ","))
+  params <- parse_name_equals_value(parse_delim(pieces[[2]], ","), ";")
 
   c(list(scheme = pieces[[1]]), params)
 }
@@ -27,7 +27,7 @@ parse_link <- function(x) {
     pieces <- parse_delim(x, ";")
 
     url <- gsub("^<|>$", "", pieces[[1]])
-    params <- parse_name_equals_value(pieces[-1])
+    params <- parse_name_equals_value(pieces[-1], ";")
     c(list(url = url), params)
   }
 
@@ -51,9 +51,14 @@ parse_delim <- function(x, delim, quote = "\"", ...) {
   )
 }
 
-parse_name_equals_value <- function(x) {
-  if (length(x) == 0) return(NULL)
-  pieces <- strsplit(x, "=")
+parse_name_equals_value <- function(x, safe = "&", quote = "\"") {
+  if (length(x) == 0) {
+    return(NULL)
+  }
+  # Note: quotes are removed in the parse_delim function
+  # re-introduce quotes to create a safe pattern to split on
+  safe_pattern <- paste0(quote, safe, quote)
+  pieces <- strsplit(sub("=", safe_pattern, x, fixed = T), safe_pattern, fixed = TRUE)
   pieces_matrix <- do.call(rbind, pieces)
 
   if (ncol(pieces_matrix) == 1) {
@@ -61,9 +66,9 @@ parse_name_equals_value <- function(x) {
   }
 
   # If only one piece, assume it's a field name with empty value
-  found <- pieces_matrix[,1] == pieces_matrix[,2]
+  found <- pieces_matrix[, 1] == pieces_matrix[, 2]
   pieces_matrix[found, 2] <- ""
-  set_names(as.list(pieces_matrix[,2]), pieces_matrix[,1])
+  set_names(as.list(pieces_matrix[, 2]), pieces_matrix[, 1])
 }
 
 parse_in_half <- function(x, pattern) {
@@ -79,6 +84,7 @@ parse_match <- function(x, pattern) {
   m <- regexec(pattern, x, perl = TRUE)
   pieces <- regmatches(x, m)[[1]][-1]
 
+  pieces <- as.list(pieces)
   # replace empty element with null
   pieces[pieces == ""] <- list(NULL)
   return(pieces)
@@ -87,4 +93,3 @@ parse_match <- function(x, pattern) {
 empty_to_null <- function(x) {
   if (x == "") NULL else x
 }
-