This is a minor performance enhancement to function `url_parse`

Issue: https://github.com/r-lib/httr2/issues/429


``` r
url <- 'https://someurl.com/with/query_string?i=main&mode=front&sid=12ab&enc=+Hello'

(bm <- bench::mark(
  httr = httr::parse_url(url),
  httr2 = httr2::url_parse(url),
  new_mthd = url_parse(url), # new method in PR
  urltools = urltools::url_parse(url),
  adaR = adaR::ada_url_parse(url),
  check = F
))
#> # A tibble: 5 × 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 httr          239µs    256µs     3774.  387.71KB     35.5
#> 2 httr2         207µs    232µs     4100.  466.62KB     34.8
#> 3 new_mthd      129µs    144µs     6509.     4.5MB     39.2
#> 4 urltools      126µs    134µs     7024.    7.45MB     43.8
#> 5 adaR          189µs    202µs     4755.    1.06MB     47.1

bm |> ggplot2::autoplot()
#> Loading required namespace: tidyr
```

![](https://i.imgur.com/kCjxbgG.png)<!-- -->

``` r

urls <- list(
  "/",
  "//google.com",
  "file:///",
  "http://google.com/",
  "http://google.com/path",
  "http://google.com/path?a=1&b=2",
  "http://google.com:80/path?a=1&b=2",
  "http://google.com:80/path?a=1&b=2#frag",
  "http://google.com:80/path?a=1&b=2&c=%7B1%7B2%7D3%7D#frag",
  "http://user@google.com:80/path?a=1&b=2",
  "http://user:pass@google.com:80/path?a=1&b=2",
  "svn+ssh://my.svn.server/repo/trunk",
  'https://someurl.com/with/query_string?i=main&mode=front&sid=12ab&enc=+Hello',
  "http://user:pass@example.com:80/path?a=1&b=2&c={1{2}3}#frag"
)

(bm <- bench::mark(
  httr = lapply(urls, httr::parse_url),
  httr2 = lapply(urls, httr2::url_parse),
  adaR = lapply(urls, adaR::ada_url_parse),
  urltools = lapply(urls, urltools::url_parse),
  new_mthd = lapply(urls, url_parse), # new method in PR
  check = F,
  iterations = 1000
))
#> # A tibble: 5 × 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 httr          2.5ms   2.55ms      388.   161.9KB     18.3
#> 2 httr2        2.25ms   2.31ms      430.    70.9KB     20.7
#> 3 adaR         2.68ms   2.78ms      358.    34.9KB     24.5
#> 4 urltools     1.73ms    1.8ms      551.    34.9KB     23.5
#> 5 new_mthd     1.59ms   1.66ms      595.    62.9KB     22.2
bm |> ggplot2::autoplot()
```

![](https://i.imgur.com/MmKgKXT.png)<!-- -->

<sup>Created on 2024-01-25 with [reprex v2.1.0](https://reprex.tidyverse.org)</sup>