@DyfanJones 2024-01-26
Before and after benchmark

```r
urls <- list(
  "/",
  "//google.com",
  "file:///",
  "http://google.com/",
  "http://google.com/path",
  "http://google.com/path?a=1&b=2",
  "http://google.com:80/path?a=1&b=2",
  "http://google.com:80/path?a=1&b=2#frag",
  "http://google.com:80/path?a=1&b=2&c=%7B1%7B2%7D3%7D#frag",
  "http://user@google.com:80/path?a=1&b=2",
  "http://user:pass@google.com:80/path?a=1&b=2",
  "svn+ssh://my.svn.server/repo/trunk",
  'https://someurl.com/with/query_string?i=main&mode=front&sid=12ab&enc=+Hello',
  "http://user:pass@example.com:80/path?a=1&b=2&c={1{2}3}#frag"
)

bm <- bench::mark(
  lapply(urls, httr2::url_parse),
  iterations = 1000
)


#> # A tibble: 2 × 13
#>   expression      min   median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc total_time result      memory              time               gc                  
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl> <int> <dbl>   <bch:tm> <list>      <list>              <list>             <list>              
#> 1 before       2.24ms   2.32ms      421.    62.9KB     46.8   900   100      2.14s <list [14]> <Rprofmem [27 × 3]> <bench_tm [1,000]> <tibble [1,000 × 3]>
#> 2 after        1.65ms   1.72ms      573.    62.9KB     20.8   965    35      1.68s <list [14]> <Rprofmem [28 × 3]> <bench_tm [1,000]> <tibble [1,000 × 3]>
```

![image](https://github.com/r-lib/httr2/assets/23235152/d692d7d5-732d-49be-9609-1a3e0c5894a9)

I am guessing there will be micro performance enchancements across httr2 as `parse_name_equals_value` is the main function that has been changed.