@DyfanJones 2024-01-29
Ah sorry about the benchmarks, my mistake:


``` r
parse_match <- function(x, pattern) {
  m <- regexec(pattern, x, perl = TRUE)
  pieces <- regmatches(x, m)[[1]][-1]
  empty <- pieces == ""
  pieces <- as.list(pieces)
  pieces[empty] <- list(NULL)
  return(pieces)
}

urls <- list(
  "/",
  "//google.com",
  "file:///",
  "http://google.com/",
  "http://google.com/path",
  "http://google.com/path?a=1&b=2",
  "http://google.com:80/path?a=1&b=2",
  "http://google.com:80/path?a=1&b=2#frag",
  "http://google.com:80/path?a=1&b=2&c=%7B1%7B2%7D3%7D#frag",
  "http://user@google.com:80/path?a=1&b=2",
  "http://user:pass@google.com:80/path?a=1&b=2",
  "svn+ssh://my.svn.server/repo/trunk",
  'https://someurl.com/with/query_string?i=main&mode=front&sid=12ab&enc=+Hello',
  "http://user:pass@example.com:80/path?a=1&b=2&c={1{2}3}#frag"
)
pattern <- "^(([^:/?#]+):)?(//([^/?#]*))?([^?#]*)(\\?([^#]*))?(#(.*))?"
bench::mark(
  before = lapply(urls, httr2:::parse_match, pattern = pattern),
  after = lapply(urls, parse_match, pattern = pattern)
) |> ggplot2::autoplot() + ggplot2::labs(title = "parse_match benchmark")
#> Loading required namespace: tidyr
```
![image](https://github.com/r-lib/httr2/assets/23235152/55e2efbc-e16f-4fc8-802f-66a154aab32d)

<sup>Created on 2024-01-29 with [reprex v2.1.0](https://reprex.tidyverse.org)</sup>

`parse_match` has a micro performance improvement.

``` r
library(rlang)
parse_name_equals_value <- function(x, safe = "&", quote = "\"") {
  if (length(x) == 0) {
    return(NULL)
  }
  # Note: quotes are removed in the parse_delim function
  # re-introduce quotes to create a safe pattern to split on
  safe_pattern <- paste0(quote, safe, quote)
  pieces <- strsplit(sub("=", safe_pattern, x, fixed = T), safe_pattern, fixed = TRUE)
  pieces_matrix <- do.call(rbind, pieces)
  
  if (ncol(pieces_matrix) == 1) {
    pieces_matrix <- cbind(pieces_matrix, rep("", nrow(pieces_matrix)))
  }
  
  # If only one piece, assume it's a field name with empty value
  found <- pieces_matrix[, 1] == pieces_matrix[, 2]
  pieces_matrix[found, 2] <- ""
  set_names(as.list(pieces_matrix[, 2]), pieces_matrix[, 1])
}

params <- "foo==&var1=12ab,var2=;var3=,&var4=\"&\"var5=;out"
bench::mark(
  before = httr2:::parse_name_equals_value(httr2:::parse_delim(params, "&")),
  after = parse_name_equals_value(httr2:::parse_delim(params, "&"), "&")
) |> ggplot2::autoplot() + ggplot2::labs(title = "parse_name_equals_value benchmark")
#> Loading required namespace: tidyr
```

![image](https://github.com/r-lib/httr2/assets/23235152/5de01334-85a3-4245-9296-0d40f48228ae)

<sup>Created on 2024-01-29 with [reprex v2.1.0](https://reprex.tidyverse.org)</sup>