@DyfanJones 2024-01-30
@hadley I have added in some extra comments and I have restructured `parse_name_equals_value` function. I am hoping it makes more sense in the logic it uses (note these changes have no impact on the current performance).